% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CUE_vec_J_bin.R
\name{CUE_vec_J_bin}
\alias{CUE_vec_J_bin}
\title{Newton Raphson Steps for plmed fitting}
\usage{
CUE_vec_J_bin(
  par,
  X,
  M,
  Y,
  Z,
  Xfam = binomial(),
  method = "G",
  weights = rep(1, N),
  med_prop = NULL,
  Sig = NULL
)
}
\arguments{
\item{par}{vector containing beta and augmented nuisance parameters, optionally with a
lagrange multiplier the final componenet.}

\item{X}{Vector of exposures}

\item{M}{Vector of mediators}

\item{Y}{Vector of outcomes}

\item{Z}{Design matrix of confounder models}

\item{method}{The method to be used. See ‘Details’.}

\item{med_prop}{Hypothesised mediation proportion for constrained optimisation.
Requires method \code{'CUE'} or \code{'Two-step'}.}

\item{Sig}{Constant covariance matrix must be prespecified when using \code{'Two-step'} method.}
}
\value{
A list containing the newton-raphson step for use with \code{\link[plmed]{newton_raph}}
as well as variance and score estimates.
}
\description{
\code{CUE_vec_J_bin} Generates newton raphson steps for the \code{\link[plmed]{plmed}}
fitting functions.
}
