% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmed.R
\name{plmed}
\alias{plmed}
\title{Partially Linear Mediation}
\usage{
plmed(
  exposure.formula,
  mediator.formula,
  outcome.formula,
  exposure.family = "binomial",
  mediator.family = "gaussian",
  data,
  weights,
  method = "G-estimation"
)
}
\arguments{
\item{exposure.formula}{an object of class "\code{\link[stats]{formula}}" (or one that can be coerced to
that class) where the left hand side of the formula contains the binary exposure variable of interest.}

\item{mediator.formula}{an object of class "\code{\link[stats]{formula}}" (or one that can be coerced to
that class) where the left hand side of the formula contains the continuous mediator variable of interest.}

\item{outcome.formula}{an object of class "\code{\link[stats]{formula}}" (or one that can be coerced to
that class) where the left hand side of the formula contains the continuous outcome variable of interest.}

\item{exposure.family}{link function for the exposure model, can be can be a character string naming a family function,
a family function or the result of a call to a family function. (See \link[stats]{family} for details of family functions.)
Must be \code{"binomial"} when using \code{Method="TTS"}. Must be \code{"none"} when using \code{Method="OLS"}.}

\item{mediator.family}{link function for the mediator model, can be either \code{"gaussian"} or \code{"binomial"}. 
Must be \code{"gaussian"} when using \code{Method="G-estimation"}}

\item{data}{an optional data frame, list or environment
(or object coercible by \code{\link{as.data.frame}}  to a data frame)
containing the variables in the model. If not found in data, the
variables are taken from environment(formula), typically the environment from
which \code{plmed} is called.}

\item{weights}{an optional vector of ‘observation weights’ to be used in 
the fitting process. Should be \code{NULL} or a numeric vector.}

\item{method}{The mediation fitting method to be used. Can be either \code{"G-estimation","TTS"} or \code{"OLS"}}
}
\value{
An object of class \code{plmed} with unconstrained parameter estimates,
 estimated standard errors, Wald based and CUE score based test statistics (G-estimation only).
}
\description{
\code{plmed} is used to fit mediation effects for a continuous outcome
given a single continuous or binary mediator, a continuous or binary exposure, and a set of
variables which adjust for confounding. This function supports three fitting methods:
those based on "\code{\link[=G_estimation]{G-estimation}}", "\code{\link{TTS}}", and "\code{\link{OLS}}".
For all methods, the confounder variable set is the union of terms in the \code{exposure.formula},
\code{mediator.formula}, and \code{outcome.formula}. Missing data behaviour is always \code{\link[stats]{na.action}=na.omit}.
}
\examples{
#Example on Generated data
N <- 100
beta <- c(1,0,1) #Some true parameter values
#Generate data on Confounders (Z), Exposure (X)
#Mediator (M), Outcome (Y)
Z <- rnorm(N)
X <- rbinom(N,1,plogis(Z))
M <- beta[1]*X + Z +rnorm(N)
Y <- beta[2]*M + beta[3]*X + Z +rnorm(N)

plmed(X~Z,M~Z,Y~Z,method="G-estimation")
plmed(X~Z,M~Z,Y~Z,method="TTS")
plmed(X~Z,M~Z,Y~Z,method="OLS")


#Example on JobsII data from the mediation package
jobs <- mediation::jobs

Z.formula = c('econ_hard','sex','age','occp',
              'marital','nonwhite','educ','income')
plmed(reformulate(Z.formula,response='treat'),
      reformulate("1",response='job_seek'),
      reformulate("1",response='depress2'),
      data=jobs)
      
#Only one of the formulas must include the confounder variables
}
