% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TTS.R
\name{TTS}
\alias{TTS}
\title{Natural Mediated Effects}
\usage{
TTS(
  exposure.formula,
  mediator.formula,
  outcome.formula,
  mediator.family = "gaussian",
  data,
  weights
)
}
\arguments{
\item{exposure.formula}{an object of class \code{\link[stats]{formula}} (or one that can be coerced to that class) 
where the left hand side of the formula contains the binary exposure variable of interest.}

\item{mediator.formula}{an object of class \code{\link[stats]{formula}} (or one that can be coerced to that class) 
where the left hand side of the formula contains the mediator variable of interest.}

\item{outcome.formula}{an object of class \code{\link[stats]{formula}} (or one that can be coerced to that class) 
where the left hand side of the formula contains the outcome variable of interest.}

\item{mediator.family}{link function for the mediator model, can be either \code{"gaussian"} or \code{"binomial"}}

\item{data}{an optional data frame, list or environment
(or object coercible by \code{\link{as.data.frame}}  to a data frame)
containing the variables in the model. If not found in data, the
variables are taken from environment(formula), typically the environment from
which \code{TTS} is called.}
}
\value{
An object of class \code{plmed} with Total Effect (TE), Natural Direct Effect (NDE) and 
Natural Indirect Effect (NIDE) estimates, with estimated standard errors, and
Wald based test statistics.
}
\description{
\code{TTS} is used to estimate natural direct and natural indirect effects using the methods of 
Tchetgen Tchetgen Shpitser (2012). This implementation can be used with a binary exposure, binary or continuous
mediator, and continuous outcome. Linear models are used to model the conditional expectation of continuous quantities,
and logisitc regression models are used to model binary quantities. As with the \code{\link{plmed}} function,
the confounder set is the union terms in the \code{exposure.formula},
\code{mediator.formula}, and \code{outcome.formula}. Missing data behaviour is always \code{\link[stats]{na.action}=na.omit}.
}
\examples{
#Example on Generated data
N <- 100
beta <- c(1,0,1) #Some true parameter values
#Generate data on Confounders (Z), Exposure (X)
#Mediator (M), Outcome (Y)
Z <- rnorm(N)
X <- rbinom(N,1,1/(exp(-Z)+1))
M <- beta[1]*X + Z +rnorm(N)
Y <- beta[2]*M + beta[3]*X + Z +rnorm(N)

TTS(X~Z,M~Z,Y~Z,"gaussian")

#Example on Generated data
N <- 100
beta <- c(1,0,1) #Some true parameter values
#Generate data on Confounders (Z), Exposure (X)
#Mediator (M), Outcome (Y)
Z <- rnorm(N)
X <- rbinom(N,1,plogis(Z))
M <- rbinom(N,1,plogis(X+Z))
Y <- beta[2]*M + beta[3]*X + Z +rnorm(N)

TTS(X~Z,M~Z,Y~Z,"binomial")
}
