% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plmed.R
\name{plmed}
\alias{plmed}
\title{Partially Linear Mediation using G-estimation}
\usage{
plmed(exposure.formula, mediator.formula, outcome.formula, data, weights)
}
\arguments{
\item{exposure.formula}{an object of class "\code{\link[stats]{formula}}" (or one that can be coerced to
that class) where the left hand side of the formula contains the binary exposure variable of interest.}

\item{mediator.formula}{an object of class "\code{\link[stats]{formula}}" (or one that can be coerced to
that class) where the left hand side of the formula contains the continuous mediator variable of interest.}

\item{outcome.formula}{an object of class "\code{\link[stats]{formula}}" (or one that can be coerced to
that class) where the left hand side of the formula contains the continuous outcome variable of interest.}

\item{data}{an optional data frame, list or environment
(or object coercible by \code{\link{as.data.frame}}  to a data frame)
containing the variables in the model. If not found in data, the
variables are taken from environment(formula), typically the environment from
which \code{plmed} is called.}

\item{weights}{an optional vector of ‘prior weights’ to be used in 
the fitting process. Should be \code{NULL} or a numeric vector.}
}
\value{
An object of class \code{plmed} with unconstrained parameter estimates,
 estimated standard errors, Wald based and CUE score based test statistics.
}
\description{
\code{plmed} is used to fit mediation effects for a continuous outcome
given a single continuous mediator, binary exposure, and set of
variables which adjust for confounding. It is assumed that the
outcome and mediator models are linear and the exposure model is
logisitic. It uses a G-estimation procedure to estimate indirect and
direct effect, with a Bias-Reduced strategy used to estimate
parameters in confounder models. Missing data behaviour is always \code{\link[stats]{na.action}=na.omit}
}
\examples{
#Example on Generated data
N <- 100
beta <- c(1,0,1) #Some true parameter values
#Generate data on Confounders (Z), Exposure (X)
#Mediator (M), Outcome (Y)
Z <- rnorm(N)
X <- rbinom(N,1,plogis(Z))
M <- beta[1]*X + Z +rnorm(N)
Y <- beta[2]*M + beta[3]*X + Z +rnorm(N)

plmed(X~Z,M~Z,Y~Z)


#Example on JobsII data from the mediation package
jobs <- mediation::jobs

Z.formula = c('econ_hard','sex','age','occp',
              'marital','nonwhite','educ','income')
plmed(reformulate(Z.formula,response='treat'),
      reformulate(Z.formula,response='job_seek'),
      reformulate(Z.formula,response='depress2'),
      data=jobs)
}
